numpy type definition consistency

+++ tsblup/simulations.py
@@ -20,7 +20,7 @@ class GeneticValueSimulator:
             "position": self.mts.sites_position[self.mts.mutations_site],
             "site_id": self.mts.mutations_site,
             "edge_id": [m.edge for m in self.mts.mutations()],
-            "effect_size": self.beta[self.mts.sites_position[self.mts.mutations_site].astype(int)],
+            "effect_size": self.beta[self.mts.sites_position[self.mts.mutations_site].astype(np.int32)],
             "trait_id": np.zeros(self.mts.num_mutations, dtype=np.int32),
             "causal_allele": [m.derived_state for m in self.mts.mutations()]
         }
